+++ SwissKnife/info/ExecutionEnvironment.py
@@ -2,40 +2,41 @@ from enum import Enum
 
 
 class ExecutionEnvironment(Enum):
-    """ExutionEnvironment is a Enum that represents a possible ExecutionEnvironment.
-    Valid values are PRO, PRE and DEV. PRE is the default value.
-    The static method "create" is the preferred method to create and ExecutionEnvironment because
-    it transform the input to an standard form (lower case) and return the default value (PRE)
-    if a invalid value (or a null value) is provided. 
+    """ExutionEnvironment is an Enum that represents a possible ExecutionEnvironment.
+    Valid values are PRO, PRE and DEV. DEV is the default value.
+    The static method "create" is the preferred method to create an ExecutionEnvironment because
+    it transforms the input to a standard form (lower case). It will return the default value (DEV)
+    if an invalid value (or a null value) is provided. 
     """
 
     PRO = "pro"
     PRE = "pre"
     DEV = "dev"
-
+    
     @staticmethod
     def create(env_str: str) -> "ExecutionEnvironment":
-        """Create an ExecutionEnvironment from the provided text. Convert it
-        convert it to "lower case" and match the result with one of the possibilities.
-        If there is not a match, PRE value will be the final result.
+        """Create an ExecutionEnvironment from the provided text. Converts it
+        to "lower case" and matches the result with one of the possibilities.
+        If there is not a match, the default value will be returned (DEV).
         
         :param env_str: Environment value in text format.
         :type env_str: str
         :return: An ExecutionEnvironment based on the previous value.
         :rtype: ExecutionEnvironment
         """
+        default_env = ExecutionEnvironment.DEV
 
         # Is none or other invalid type.
         if type(env_str) is not str:
-            return ExecutionEnvironment("pre")
+            return default_env
 
         try:
             return ExecutionEnvironment(env_str.lower())
         except ValueError:
-            return ExecutionEnvironment("pre")
+            return default_env
 
     def is_pro(self):
-        """Return if the value of this ExecutionEnvironment is PRO.
+        """Indicates if the value of this ExecutionEnvironment is PRO.
         
         :return: True if this ExecutionEnvironment is PRO.
         :rtype: bool
@@ -43,7 +44,7 @@ class ExecutionEnvironment(Enum):
         return self == ExecutionEnvironment.PRO
 
     def is_pre(self):
-        """Return if the value of this ExecutionEnvironment is PRE.
+        """Indicates if the value of this ExecutionEnvironment is PRE.
         
         :return: True if this ExecutionEnvironment is PRE.
         :rtype: bool
@@ -52,7 +53,7 @@ class ExecutionEnvironment(Enum):
         return self == ExecutionEnvironment.PRE
 
     def is_dev(self):
-        """Return if the value of this ExecutionEnvironment is DEV.
+        """Indicates if the value of this ExecutionEnvironment is DEV.
         
         :return: True if this ExecutionEnvironment is DEV.
         :rtype: bool