+++ tests/info/test_ExecutionEnvironment.py
@@ -26,10 +26,7 @@ class TestExecutionEnvironment(unittest.TestCase):
 
         for env_value in ["asdfasdf", "PRU", "Der","",None]:
             execution_environment = ExecutionEnvironment.create(env_value)
-            self._only_is_pre(execution_environment)
-
-    
-   # def test_current_environment_object(self):
+            self._only_is_dev(execution_environment)
 
     def _only_is_pro(self, execution_environment: ExecutionEnvironment):