+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.1.1] - 2019-10-31
+### Added
+- Specific version of the project dependencies.
+  
+### Changed
+- `ExecutionEnvironment` default value set to **DEV**.
+- Fixed some typos in the code.
+  
 
 ## [0.1.0] - 2019-10-31
 ### Added